requirements.txt
pika==1.3.1
python-dotenv==0.21.1
selenium==4.8.2
setuptools==45.2.0
undetected_chromedriver==2.1.1
